requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioice==0.10.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
audioread==3.0.1
av==13.1.0
aws-sdk-signers==0.0.3
aws_sdk_bedrock_runtime==0.0.2
awscrt==0.27.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
cryptography==45.0.4
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
fastapi==0.115.13
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.27.0
google-crc32c==1.7.1
google-genai==1.21.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
h264-profile-level-id==1.0.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
ijson==3.4.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.4
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.0
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.0
openai==1.88.0
packaging==25.0
pillow==10.4.0
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyee==11.1.0
Pygments==2.19.1
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
sarvamai==0.1.5
scikit-learn==1.7.0
scipy==1.15.3
sdp-transform==1.1.0
smithy-aws-core==0.0.3
smithy-aws-event-stream==0.0.1
smithy-core==0.0.2
smithy-http==0.0.1
smithy-json==0.0.1
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sse-starlette==2.3.6
starlette==0.46.2
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
videosdk==0.1.0
videosdk-agents==0.0.14
videosdk-plugins-aws==0.0.5
videosdk-plugins-deepgram==0.0.3
videosdk-plugins-elevenlabs==0.0.2
videosdk-plugins-google==0.0.6
videosdk-plugins-openai==0.0.8
videosdk-plugins-sarvamai==0.0.1
videosdk-plugins-silero==0.0.2
videosdk-plugins-turn-detector==0.0.3
vsaiortc==0.0.8
websockets==15.0.1
yarl==1.20.1